requirements.txt
requests==2.33.1
beautifulsoup4==4.14.3
langdetect==1.0.9
matplotlib==3.10.8
nltk==3.9.4
pandas==3.0.2
seaborn==0.13.2
vaderSentiment==3.3.2
streamlit==1.56.0
plotly==6.7.0
networkx==3.6.1
scipy==1.15.2
PyYAML==6.0.2
